ZOMBIS CATCHER AND DATA-STRUCTURE DESTRUCTOR SUSCESSFULLY IMPLEMENTED (just causes litle issues with the prompt)

+++ srcs/builtins/bg.c
@@ -6,7 +6,7 @@
 /*   By: pablo <pablo@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/11/18 23:11:42 by pablo             #+#    #+#             */
-/*   Updated: 2020/11/23 09:14:43 by pablo            ###   ########.fr       */
+/*   Updated: 2020/11/23 14:28:11 by pablo            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -38,12 +38,15 @@ void	resume_background_group(t_process* leader)
 			remember_leader = g_session->groups->active_processes;
 			while (g_session->groups->active_processes != g_session->groups->nil)
 			{
-				ft_dprintf(2, "[BG][KILL -SIGCONT \'%d\'][\'%p\']\n", g_session->groups->active_processes->pid, g_session->groups->active_processes);
-				g_session->groups->active_processes->flags &= ~STOPPED;
-				g_session->groups->active_processes->flags |= BACKGROUND;
-				kill(g_session->groups->active_processes->pid, SIGCONT);
-				ft_dprintf(2, "[BG][TARGET FLAGS: %d][\'%p\']\n", g_session->groups->active_processes->flags, g_session->groups->active_processes);
-				//update_background(g_session, &g_session->groups->active_processes, true);
+				if (g_session->groups->active_processes->flags & STOPPED)
+				{
+					ft_dprintf(2, "[BG][KILL -SIGCONT \'%d\'][\'%p\']\n", g_session->groups->active_processes->pid, g_session->groups->active_processes);
+					g_session->groups->active_processes->flags &= ~STOPPED;
+					g_session->groups->active_processes->flags |= BACKGROUND;
+					kill(g_session->groups->active_processes->pid, SIGCONT);
+					ft_dprintf(2, "[BG][TARGET FLAGS: %d][\'%p\']\n", g_session->groups->active_processes->flags, g_session->groups->active_processes);
+					//update_background(g_session, &g_session->groups->active_processes, true);
+				}
 				g_session->groups->active_processes = g_session->groups->active_processes->next;
 			}
 			/*
@@ -83,7 +86,7 @@ int		ft_bg(t_exec* args, t_term* term)
     }
 	t_group*	remember;
 	remember = g_session->groups;
-	while (g_session->groups != g_session->nil)
+	while (g_session->groups != g_session->nil->prev)
 	{
     	target = g_session->groups->active_processes == g_session->groups->nil ? &g_session->groups->next->active_processes : &g_session->groups->active_processes;
 		if ((*target)->flags & BACKGROUND)
@@ -105,6 +108,11 @@ int		ft_bg(t_exec* args, t_term* term)
 	{
 		ft_dprintf(STDERR_FILENO, "minish: bg: %s: no such job\n", args->av[1]);
             return (STD_ERROR);
+	}
+	if ((*target)->flags & BACKGROUND)
+	{
+		ft_dprintf(STDERR_FILENO, "minish: job %lu already in background\n", get_background_index(g_session->nil, *target));
+		return (SUCCESS);
 	}
 	//ft_dprintf(2, "target = %p\n", target);
 	//ft_dprintf(2, "*target = %p\n", *target);
